chore(`ffi`): split into modules
